refactor: colocate DeepEncoder Triton attention
